requirements.txt
aiohttp>=3.8.4
fastapi>=0.104.1
numpy>=1.24.3
obs-websocket-py>=1.0
Pillow>=10.0.0
uvicorn>=0.23.2
websockets>=11.0.3
starlette>=0.27.0
pydantic>=2.0.0